SimuladorRedeCan/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-can==4.3.1